Credit creator and method for jollof prep
I realized in the last version, the author of the recipe wasn't
credited so I put her name there to indicate this isn't my recipe.
I also added the ingredients and the steps for the rice. There is no
recipe without knowing what you need and what order to do things in.